move requests, hooks and types to corresponding folders remove unnecessary functions and variables
