+++ idea/frontend/.env.example
@@ -1,7 +1,5 @@
 VITE_API_URL=
 VITE_INDEXER_API_URL=
-VITE_MAINNET_VOUCHERS_API_URL=
-VITE_TESTNET_VOUCHERS_API_URL=
 VITE_NODES_API_URL=
 VITE_NODE_ADDRESS=
 VITE_HCAPTCHA_SITE_KEY=
