+++ idea/frontend/src/features/voucher/api/consts.ts
@@ -1,0 +1,6 @@
+const METHOD = {
+  DATA: 'voucher.data',
+  ALL: 'voucher.all',
+} as const;
+
+export { METHOD };
