+++ idea/frontend/src/features/voucher/api/index.ts
@@ -1,0 +1,4 @@
+import { getVoucher, getVouchers } from './requests';
+import { METHOD } from './consts';
+
+export { getVoucher, getVouchers, METHOD };
