+++ idea/frontend/src/features/voucher/hooks/index.ts
@@ -1,6 +1,6 @@
 import { useDurationSchema } from './use-duration-schema';
 import { useVoucherType } from './use-voucher-type';
-import { useVouchers } from './use-vouchers';
+import { useVouchers } from '../api/hooks/use-vouchers';
 import { useVoucherFilters } from './use-voucher-filters';
 
 export { useDurationSchema, useVoucherType, useVouchers, useVoucherFilters };
