+++ idea/frontend/src/features/voucher/types.ts
@@ -9,37 +9,4 @@ type Values = {
   [FIELD_NAME.DURATION]: string;
 };
 
-type Voucher = {
-  id: HexString;
-  owner: HexString;
-  spender: HexString;
-  amount: string;
-  balance: string;
-  programs: HexString[];
-  codeUploading: boolean;
-  expiryAtBlock: string;
-  expiryAt: string;
-  issuedAtBlock: string;
-  issuedAt: string;
-  updatedAtBlock: string;
-  updatedAt: string;
-  isDeclined: boolean;
-};
-
-type VouchersParams = {
-  limit: number;
-  offset: number;
-  id: HexString;
-  owner: string;
-  spender: string;
-  programs: HexString[];
-  declined: boolean;
-  expired: boolean;
-};
-
-type VouchersResponse = {
-  vouchers: Voucher[];
-  count: number;
-};
-
-export type { VoucherType, Values, Voucher, VouchersParams, VouchersResponse };
+export type { VoucherType, Values };
