+++ idea/frontend/src/features/voucher/ui/update-voucher-modal/update-voucher-modal.tsx
@@ -15,10 +15,11 @@ import { ACCOUNT_ADDRESS_SCHEMA } from '@/shared/config';
 
 import { DEFAULT_VALUES, FIELD_NAME } from '../../consts';
 import { useDurationSchema } from '../../hooks';
-import { Values, Voucher } from '../../types';
+import { Values } from '../../types';
 import { DurationForm } from '../duration-form';
 import { ProgramsForm } from '../programs-form';
 import styles from './update-voucher-modal.module.scss';
+import { Voucher } from '../../api/types';
 
 type Props = {
   voucher: Voucher;
