+++ idea/frontend/src/features/voucher/ui/update-voucher/update-voucher.tsx
@@ -3,8 +3,8 @@ import { Button } from '@gear-js/ui';
 import EditSVG from '@/shared/assets/images/actions/edit.svg?react';
 import { useModalState } from '@/hooks';
 import { UpdateVoucherModal } from '../update-voucher-modal';
-import { Voucher } from '../../types';
 import styles from './update-voucher.module.scss';
+import { Voucher } from '../../api/types';
 
 type Props = {
   voucher: Voucher;
