+++ idea/frontend/src/features/voucher/utils.ts
@@ -1,7 +1,4 @@
 import { DEFAULT_LIMIT } from '@/shared/config';
-import { fetchWithGuard } from '@/shared/helpers';
-
-import { VouchersResponse } from './types';
 
 const MULTIPLIER = { MS: 1000, S: 60, M: 60, H: 24, D: 30 };
 
@@ -33,20 +30,4 @@ function getTime(ms: number) {
   return result.trim();
 }
 
-const getVouchers = (url: string, params: { programs?: string[]; offset: number; limit: number; id: string }) => {
-  const method = 'POST';
-  const body = JSON.stringify(params);
-  const headers = { 'Content-Type': 'application/json' };
-
-  return fetchWithGuard<VouchersResponse>(`${url}/vouchers`, { method, body, headers });
-};
-
-const getNextPageParam = (lastPage: VouchersResponse, allPages: VouchersResponse[]) => {
-  const totalCount = lastPage.count;
-  const lastPageCount = lastPage.vouchers.length;
-  const fetchedCount = (allPages.length - 1) * DEFAULT_LIMIT + lastPageCount;
-
-  return fetchedCount < totalCount ? fetchedCount : undefined;
-};
-
-export { getPluralizedUnit, getMilliseconds, getTime, getVouchers, getNextPageParam };
+export { getPluralizedUnit, getMilliseconds, getTime };
